FEATURE: switch volume to be in Long type

+++ src/main/java/com/citi/group77777/model/StockIndex.java
@@ -116,7 +116,7 @@ public class StockIndex {
         return volume;
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(long volume) {
         this.volume = volume;
     }
 }
